Manage a seller's tax registrations the way a company's are
A seller is a business like any other: one trading in several regimes holds a
registration in each, and they are not alternatives. The panel offered a single
VAT field on the profile form, which meant a seller could record only one and
the surface disagreed with the `has_many` the model has always declared.

Registrations move to their own collection — `/seller/tax_identifiers`, the
same shape a company's take, including the re-check action — and off the
profile, whose single field is gone.

The panel itself is now one component in `@spree/dashboard-ui`, rendered by
both dashboards. Two components for one job is how the hardcoded kind and the
toast hidden behind the sheet came to exist in only one of them. It stays
headless: the host passes its own mutations and the error mapper that knows
which SDK raised a failure.

Re-run CI

The test workflows did not fire for ca1053e, so its checks were green only
because nothing ran them.

Carry the rebase through the generated files

Resolving the conflict by hand took a stale copy of the seller SDK type and
the OpenAPI spec, dropping the policies field and the password-reset endpoints
main added. Both are regenerated from the serializers instead, and the export
params type gets back the closing brace the resolution ate.
