+++ packages/dashboard/src/locales/ar.json
@@ -1990,11 +1990,13 @@
       "tax_identifier": {
         "kind": {
           "help": "نوع هذا الرقم. يجب أن يطابق التحقق المُعد له، لذا فالكتابة الصحيحة مهمة.",
-          "label": "النوع"
+          "label": "النوع",
+          "required": "اختر نوع التسجيل أو اكتبه"
         },
         "value": {
           "label": "الرقم",
-          "placeholder": "DE123456789"
+          "placeholder": "DE123456789",
+          "required": "أدخل رقم التسجيل"
         }
       },
       "tax_rate": {
