+++ packages/dashboard/src/locales/de.json
@@ -1976,11 +1976,13 @@
       "tax_identifier": {
         "kind": {
           "help": "Um welche Nummer es sich handelt. Sie muss zur eingerichteten Prüfung passen, die Schreibweise ist also wichtig.",
-          "label": "Art"
+          "label": "Art",
+          "required": "Registrierungsart wählen oder eingeben"
         },
         "value": {
           "label": "Nummer",
-          "placeholder": "DE123456789"
+          "placeholder": "DE123456789",
+          "required": "Registrierungsnummer eingeben"
         }
       },
       "tax_rate": {
