+++ packages/dashboard/src/locales/pl.json
@@ -1988,11 +1988,13 @@
       "tax_identifier": {
         "kind": {
           "help": "Jaki to numer. Musi odpowiadać skonfigurowanej weryfikacji, więc pisownia ma znaczenie.",
-          "label": "Rodzaj"
+          "label": "Rodzaj",
+          "required": "Wybierz lub wpisz rodzaj rejestracji"
         },
         "value": {
           "label": "Numer",
-          "placeholder": "PL1234567890"
+          "placeholder": "PL1234567890",
+          "required": "Podaj numer rejestracyjny"
         }
       },
       "tax_rate": {
