+++ packages/seller-dashboard/src/pages/profile.tsx
@@ -35,6 +35,7 @@ import { CenteredMessage } from '../components/centered-message'
 import { SellerAddressCard } from '../components/seller-address-card'
 import { SellerBusinessCard } from '../components/seller-business-card'
 import { SellerReturnsLocationCard } from '../components/seller-returns-location-card'
+import { SellerTaxIdentifiersCard } from '../components/seller-tax-identifiers-card'
 
 interface ProfileValues {
   name: string
@@ -90,6 +91,7 @@ export function ProfilePage() {
             <AtAGlanceCard profile={profile} teamCount={team?.data?.length} />
             <ContactCard profile={profile} onEdit={() => setEditing(true)} />
             <SellerBusinessCard profile={profile} />
+            <SellerTaxIdentifiersCard />
             <SellerAddressCard profile={profile} />
             <SellerReturnsLocationCard />
             <SettlementCard profile={profile} />
