+++ spree/api/app/controllers/concerns/spree/api/v3/tax_identifier_validation.rb
@@ -1,0 +1,41 @@
+module Spree
+  module Api
+    module V3
+      # The re-check action, shared by every surface where a registration is
+      # managed: nested under a customer, under a company, and the seller's
+      # own panel.
+      #
+      # Separate from the number-changed path, which the model stamps and a
+      # subscriber queues. This is staff asking "is it still valid now" — a
+      # number verified last year may have been deregistered since — so it
+      # queues a check without the number having changed.
+      module TaxIdentifierValidation
+        extend ActiveSupport::Concern
+
+        def validate
+          authorize_resource!(@resource, :update)
+
+          # A kind can be format-checked here and still have no registry to
+          # ask — core ships exactly that for EU VAT — so the message says
+          # what is missing rather than claiming nothing is registered.
+          unless @resource.validatable?
+            render_error(
+              code: 'tax_id_not_validatable',
+              message: "No registry validator is installed for tax identifier kind '#{@resource.kind}'",
+              status: :unprocessable_content
+            )
+            return
+          end
+
+          # Marked pending before the job is queued, not after: the job can
+          # finish first, and writing 'pending' afterwards would bury the
+          # verdict it just recorded.
+          @resource.update_columns(validation_status: 'pending', validated_at: nil, updated_at: Time.current)
+          Spree::TaxIdentifiers::ValidateJob.perform_later(@resource.id)
+
+          render json: serialize_resource(@resource.reload), status: :accepted
+        end
+      end
+    end
+  end
+end
