+++ spree/api/spec/controllers/spree/api/v3/seller/profile_controller_spec.rb
@@ -228,86 +228,6 @@ RSpec.describe Spree::Api::V3::Seller::ProfileController, type: :controller do
       expect(seller.registration_number).to eq('01234567')
     end
 
-    let(:vat_number) { eu_vat_number(0) }
-    let(:corrected_vat_number) { eu_vat_number(1) }
-
-    it 'records a tax registration' do
-      patch :update, params: { tax_identifier: { kind: 'eu_vat', value: vat_number } }, as: :json
-
-      expect(response).to have_http_status(:ok)
-      identifier = seller.reload.tax_identifiers.find_by(kind: 'eu_vat')
-      expect(identifier.value).to eq(vat_number)
-    end
-
-    # One per kind, so correcting the number replaces it rather than stacking
-    # a second row the model would refuse.
-    it 'corrects a registration rather than adding a second' do
-      patch :update, params: { tax_identifier: { kind: 'eu_vat', value: vat_number } }, as: :json
-      patch :update, params: { tax_identifier: { kind: 'eu_vat', value: corrected_vat_number } }, as: :json
-
-      expect(response).to have_http_status(:ok)
-      identifiers = seller.reload.tax_identifiers.where(kind: 'eu_vat')
-      expect(identifiers.count).to eq(1)
-      expect(identifiers.first.value).to eq(corrected_vat_number)
-    end
-
-    # The seller panel lets the kind change, so a seller moving from one regime
-    # to another must end up holding the new registration and not both.
-    it 'replaces the registration when the seller changes its kind' do
-      patch :update, params: { tax_identifier: { kind: 'eu_vat', value: vat_number } }, as: :json
-      patch :update, params: { tax_identifier: { kind: 'gb_vat', value: 'GB123456789' } }, as: :json
-
-      expect(response).to have_http_status(:ok)
-      identifiers = seller.reload.tax_identifiers
-      expect(identifiers.map(&:kind)).to contain_exactly('gb_vat')
-      expect(identifiers.first.value).to eq('GB123456789')
-    end
-
-    # Silently dropping it would answer 200 to a number that was never stored.
-    it 'refuses a malformed number rather than discarding it quietly' do
-      patch :update, params: { tax_identifier: { kind: 'eu_vat', value: 'DE123' } }, as: :json
-
-      expect(response).to have_http_status(:unprocessable_content)
-      expect(seller.reload.tax_identifiers.where(kind: 'eu_vat')).to be_empty
-    end
-
-    # A 422 the seller reads as "nothing saved" must mean it.
-    it 'keeps the rest of the form out of the database when the number is refused' do
-      original_name = seller.name
-
-      patch :update, params: { name: 'Renamed Seller',
-                               tax_identifier: { kind: 'eu_vat', value: 'DE123' } }, as: :json
-
-      expect(response).to have_http_status(:unprocessable_content)
-      expect(seller.reload.name).to eq(original_name)
-    end
-
-    # A stale form field must not erase a registration the seller never touched.
-    it 'clears only the kind the blank was sent for' do
-      seller.tax_identifiers.create!(kind: 'gb_vat', value: 'GB123456789')
-
-      patch :update, params: { tax_identifier: { kind: 'eu_vat', value: '' } }, as: :json
-
-      expect(response).to have_http_status(:ok)
-      expect(seller.reload.tax_identifiers.map(&:kind)).to contain_exactly('gb_vat')
-    end
-
-    it 'removes it when the seller clears the number' do
-      patch :update, params: { tax_identifier: { kind: 'eu_vat', value: vat_number } }, as: :json
-      patch :update, params: { tax_identifier: { kind: 'eu_vat', value: '' } }, as: :json
-
-      expect(seller.reload.tax_identifiers.where(kind: 'eu_vat')).to be_empty
-    end
-
-    it 'serializes them back' do
-      seller.tax_identifiers.create!(kind: 'eu_vat', value: vat_number)
-
-      get :show, as: :json
-
-      expect(json_response['tax_identifiers'].first).to include(
-        'kind' => 'eu_vat', 'value' => vat_number
-      )
-    end
   end
 
   # The profile inherits its policies association from the store serializer,
